Remove duplicate alerts tests from T1
In order to reduce the observability lane execution time,
I removed alerts tests that has coverage in T1.
In addition added coverage for the metrics:
kubevirt_cdi_operator_up
kubevirt_virt_api_up
kubevirt_virt_controller_up
