+++ tests/install_upgrade_operators/product_upgrade/test_upgrade_iuo.py
@@ -35,14 +35,14 @@ class TestUpgradeIUO:
         prometheus_scope_function,
         fired_alerts_during_upgrade,
     ):
-        LOGGER.info("Verify if any alerts were fired during upgrade")
+        LOGGER.info("Verify if any runbook_url were fired during upgrade")
         process_alerts_fired_during_upgrade(
             prometheus=prometheus_scope_function,
             fired_alerts_during_upgrade=fired_alerts_during_upgrade,
         )
         if fired_alerts_during_upgrade:
             collect_alerts_data()
-            raise AssertionError(f"Following alerts were fired during upgrade: {fired_alerts_during_upgrade}")
+            raise AssertionError(f"Following runbook_url were fired during upgrade: {fired_alerts_during_upgrade}")
 
     @pytest.mark.polarion("CNV-6866")
     @pytest.mark.order(before=IMAGE_UPDATE_AFTER_UPGRADE_NODE_ID, after=IUO_CNV_ALERT_ORDERING_NODE_ID)
