+++ tests/observability/metrics/test_cdi_metrics.py
@@ -1,6 +1,7 @@
 import pytest
 
 from tests.observability.utils import validate_metrics_value
+from utilities.constants import CDI_OPERATOR
 
 
 @pytest.mark.polarion("CNV-10557")
@@ -28,4 +29,27 @@ def test_kubevirt_cdi_upload_pods_high_restart(
         prometheus=prometheus,
         expected_value="1",
         metric_name="kubevirt_cdi_upload_pods_high_restart",
+    )
+
+
+@pytest.mark.parametrize(
+    "scaled_deployment",
+    [
+        pytest.param(
+            {"deployment_name": CDI_OPERATOR, "replicas": 0},
+            marks=(pytest.mark.polarion("CNV-11722")),
+            id="Test_kubevirt_cdi_operator_up",
+        ),
+    ],
+    indirect=True,
+)
+def test_kubevirt_cdi_operator_up(
+    prometheus,
+    disabled_virt_operator,
+    scaled_deployment,
+):
+    validate_metrics_value(
+        prometheus=prometheus,
+        expected_value="0",
+        metric_name="kubevirt_cdi_operator_up",
     )
