--- tests/observability/storage/constants.py
@@ -1,1 +1,0 @@
-HPP_NOT_READY = "HPPNotReady"
